[api]: Implement Watney plate solver

+++ nebulosa-astap/src/main/kotlin/nebulosa/astap/star/detection/Star.kt
@@ -3,8 +3,8 @@ package nebulosa.astap.star.detection
 import nebulosa.star.detection.ImageStar
 
 data class Star(
-    override val x: Int = 0,
-    override val y: Int = 0,
+    override val x: Double = 0.0,
+    override val y: Double = 0.0,
     override val hfd: Double = 0.0,
     override val snr: Double = 0.0,
     override val flux: Double = 0.0,
