+++ nebulosa-watney/src/main/kotlin/nebulosa/watney/star/detection/Star.kt
@@ -3,7 +3,7 @@ package nebulosa.watney.star.detection
 import nebulosa.star.detection.ImageStar
 
 data class Star(
-    override val x: Int = 0, override val y: Int = 0,
+    override val x: Double = 0.0, override val y: Double = 0.0,
     val size: Double = 0.0,
     override var hfd: Double = 0.0,
     override var snr: Double = 0.0,
